refactor: move and rename lists from lib dir in settings
