+++ src/lib/settings/settingsList.js
@@ -1,8 +1,12 @@
-import { chatbot } from "./botDetails";
+import { chatbot } from "../botDetails";
 
 export const settingsList = [
   {
     name: "appearance",
+    settingType: "chat",
+  },
+  {
+    name: "sound",
     settingType: "chat",
   },
   {
